+++ pages/document.py
@@ -1,11 +1,11 @@
-from utilities.text_processor import TextProcessor
+from utilities.document_processor import TextProcessor
 import streamlit as st
 
 # Prompt for LLM
 prompt = """
-You are a technical writer tasked with creating a user guide for a product/service. 
-Your task is to create a user guide by incorporating relevant information from the provided outdated user guide, and the reference material.
-Only generate the content that is to be used in this new user guide.
+You are a technical writer tasked with updating an outdated user guide for a product/service. 
+Your task is to update an outdated user guide by incorporating relevant information from the provided reference material.
+Only generate the content that is to be used in this updated user guide.
 Do not provide explanations or notes.
 
 The user will input one text document
@@ -14,13 +14,13 @@ Document 2: Reference Material
 """
 
 def main():
-    st.set_page_config(page_title="User Guide Creator", layout="wide")
+    st.set_page_config(page_title="Update a Document", layout="wide")
 
     # Set app header
-    st.header("Create a user guide")
+    st.header("Update a Document")
 
     # Upload both outdated user guide and reference material
-    outdated_guide_file = st.file_uploader("Choose the outdated user guide document", type=["txt", "pdf", "docx", "pptx"])
+    outdated_guide_file = st.file_uploader("Choose the outdated user guide document", type=["txt", "pdf", "docx"])
     reference_material_file = st.file_uploader("Choose the reference material document", type=["txt", "pdf", "docx", "pptx"])
     webpage_link = st.text_input("Input webpage link")
 
@@ -44,16 +44,16 @@ def main():
                     st.session_state.generated_text = TextProcessor.generate_text(prompt, st.session_state.outdated_guide_text,
                                                                                   st.session_state.reference_material_text)
                     #st.write(reference_material_text)
-
             else:
                 st.error("Please upload either reference material or webpage link.")
 
         else:
             st.error("Please upload outdated material.")
 
-    st.subheader("Updated User Guide Content")
+
 
     if st.session_state.generated_text:
+        st.subheader("Updated User Guide Content")
 
         if 'edit_mode' not in st.session_state:
             st.session_state.edit_mode = False
@@ -71,12 +71,12 @@ def main():
 
         st.button("Re-Generate", on_click=regenerate_guide)
 
-        TextProcessor.select_version()
+        #TextProcessor.select_version()
 
         st.subheader("Download the Updated User Guide")
 
-        st.selectbox("Export As...",
-                     ("DOCX", "TXT", "PDF", "PPT", "MP4"))
+        # st.selectbox("Export As...",
+        #              ("DOCX", "TXT", "PDF", "PPT", "MP4"))
 
         st.download_button(
             "Download",
